refactoring: 언팔로우 수정
언팔로우 수정

 #34
